添加PID代码

+++ Apply/Task/Inc/task_motionmode.h
@@ -3,7 +3,7 @@
 
 #define pi 3.1415926
 
-float *Task_Horizontal_Heading_Calculation(float P,int θ);
+float *Task_Horizontal_Heading_Calculation(uint16_t P,int θ);
 
 #endif
